refactor: goal label
